refactor error handling in the aggregator
Summary and motivation
======================

In order to return proper errors to the clients in the API layer, we
need errors to be reported accurately within the service: a mere
`Result<T, ()>` is not good enough. This commit refactors the error
handling in the aggregator service, such that any error (whether it
happens in `PyAggregator`, in the RPC layer, or in the protocol
itself) is propagated back to the `ServiceHandle`, which is the
entrypoint for making requests to the service.

Implementation
==============

`ServiceError<T>`
-----------------

The `ServiceHandle` methods now return a `ServiceError<T>`, where `T`
depends on which method is called. This way, `ServiceError` can
describe both errors that are common to all the methods, and errors
that are specific to certain methods. Thus the `ServiceHandle` now
looks like this:

```rust
impl<A> ServiceHandle<A>
where
  A: Aggregator + 'static,
{
  async fn download(&self, credentials: Credentials) -> Result<Bytes, ServiceError<DownloadError>> {
    // ...
  }
  async fn upload(&self, credentials: Credentials, data: Bytes) -> Result<(), ServiceError<UploadError>> {
    // ...
  }
  async fn aggregate(&self) -> Result<(), ServiceError<A::Error>> {
    // ...
  }
  async fn select(&self, credentials: Credentials) -> Result<(), ServiceError<A::Error>> {
    // ...
  }
  // ...
}

Aggregator's error
------------------

Note that `aggregate` and `select`'s error type is tied to the
`Aggregator` associated type `Aggregator::Error`. This is because we
don't want to force implementors of `Aggregator` to use a certain
error type. The downside is that is makes `ServiceHandle` generic over
`A: Aggregator`. We now have:

```rust
pub struct ServiceHandle<A: Aggregator> {
  // ..
}
```

Instead of

```rust
pub struct ServiceHandle {
  // ..
}
```

We could maybe simplify this buy defining `Aggregator` as:

```rust
pub trait Aggregator {
    type Error: Error + Send + 'static + Sync + Into<AggregatorError>;
    type AggregateFut: Future<Output = Result<Bytes, Self::Error>> + Unpin;
    type AddWeightsFut: Future<Output = Result<(), Self::Error>> + Unpin + Send + 'static;
```

where `AggregatorError` is some error that we define. But for now, we
only have one `Aggregator` implementation so it is difficult to
imagine what `AggregatorError` should look like.

For now, our only `Aggregator` implementation, `PyAggregatorHandle`,
uses the following error type:

```rust
pub enum PyAggregatorHandleError {
    #[error("failed to send the request or receive the response")]
    Handle(#[from] ChannelError),

    #[error("error while executing Python code")]
    Python(#[from] PythonError),

    #[error("request failed: {0}")]
    Request(String),
}
```

RPC error
---------

On the RPC side, the error handling was also improved. The improvement
is twofolds:

- The RPC server now returns `Result<T, ServerError>` instead of
  `Result<T, ()>`
- The RPC client returns simpler `Result<T, ClientError>` instead of
  nested results `io::Result<Result<T, ()>>`. The `ClientError` simply
  wraps the `io::Error` and `ServerError` cases.